removed .vscode

+++ internal/time.go
@@ -5,5 +5,5 @@ import "time"
 const Year = time.Hour * 24 * 365
 
 func TimesInYear(duration time.Duration) float64 {
-	return float64(Year / duration)
+	return float64(Year) / float64(duration)
 }
